add and modified logic

+++ modules/filemanager/controllers/AlbumController.php
@@ -7,13 +7,16 @@ use app\modules\dao\ar\Taxonomy;
 use app\modules\filemanager\searchs\AlbumSearc;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
+use yii\web\HttpException;
 use yii\filters\VerbFilter;
 use app\modules\filemanager\FileManager;
+
 /**
  * AlbumController implements the CRUD actions for Taxonomy model.
  */
 class AlbumController extends Controller
 {
+
     public function behaviors()
     {
         return [
@@ -32,13 +35,17 @@ class AlbumController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new AlbumSearc;
-        $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
+        if (Yii::$app->user->can('albumindex')) {
+            $searchModel = new AlbumSearc;
+            $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-            'searchModel' => $searchModel,
-        ]);
+            return $this->render('index', [
+                        'dataProvider' => $dataProvider,
+                        'searchModel' => $searchModel,
+            ]);
+        } else {
+            throw new HttpException('The requested page does not exist.');
+        }
     }
 
     /**
@@ -48,9 +55,13 @@ class AlbumController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if (Yii::$app->user->can('albumview')) {
+            return $this->render('view', [
+                        'model' => $this->findModel($id),
+            ]);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
@@ -60,16 +71,20 @@ class AlbumController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Taxonomy;
-        $model->setAttribute('term_id', FileManager::FILE_IMAGE_TERM);
-        $model->setAttribute('create_et', date("Y-m-d H:i:s"));
-        $model->setAttribute('update_et', date("Y-m-d H:i:s"));
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        if (Yii::$app->user->can('albumcreate')) {
+            $model = new Taxonomy;
+            $model->setAttribute('term_id', FileManager::FILE_IMAGE_TERM);
+            $model->setAttribute('create_et', date("Y-m-d H:i:s"));
+            $model->setAttribute('update_et', date("Y-m-d H:i:s"));
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('create', [
+                            'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
@@ -81,14 +96,18 @@ class AlbumController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (Yii::$app->user->can('albumupdate')) {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('update', [
+                            'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
@@ -100,9 +119,13 @@ class AlbumController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->can('albumdelete')) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
@@ -120,4 +143,5 @@ class AlbumController extends Controller
             throw new NotFoundHttpException('The requested page does not exist.');
         }
     }
+
 }
