+++ modules/filemanager/controllers/DocumentController.php
@@ -15,6 +15,7 @@ use Symfony\Component\Filesystem\Filesystem;
 use app\modules\filemanager\FileManager;
 use yii\web\UploadedFile;
 use app\modules\filemanager\models\DocumentModel;
+use yii\web\HttpException;
 
 /**
  * DocumentController implements the CRUD actions for File model.
@@ -47,80 +48,86 @@ class DocumentController extends Controller
     public function actionUploaddredactorfile()
     {
 //        echo date('Ym',strtotime("2009-05-23 01:25:08"));
+        if (Yii::$app->user->can('documentuploaddredactorfile')) {
+            $file = UploadedFile::getInstanceByName('file');
+            if ($file->size === 0) {
+                echo Json::encode([
+                    'error' => 'Ukuran file yang anda unggah terlalu besar! Maksimum ' . ini_get('upload_max_filesize')
+                ]);
+            } else {
+                $dir = date("Ym");
+                $this->filesystem->mkdir($this->path . $dir);
+                $unique_name = preg_replace('/\s+/', '-', date("YmdHis") . '-' . rand(0, 999999999) . '-' . $file->name);
 
-        $file = UploadedFile::getInstanceByName('file');
-        if ($file->size === 0) {
-            echo Json::encode([
-                'error'=>'Ukuran file yang anda unggah terlalu besar! Maksimum ' . ini_get('upload_max_filesize')
-             ]);
+                $file->saveAs($this->path . $dir . DS . $unique_name);
+
+                /**
+                 * Save DocumentModel name and order attr to database
+                 */
+                $model = new DocumentModel();
+                $model->user_id = Yii::$app->user->identity->getId();
+                $model->name = $file->name;
+                $model->orginal_name = $file->name;
+                $model->unique_name = $unique_name;
+                $model->type = FileManager::FILE_TYPE_DOCUMENT;
+                $model->size = "$file->size";
+                $model->file_type = $file->type;
+                $model->description = "Dokumen " . $file->name;
+                $model->create_at = date("Y-m-d H:i:s");
+                $model->update_et = date("Y-m-d H:i:s");
+                $model->save();
+                $base_path = Yii::getAlias('@web') . DS . 'resources' . DS . 'documents' . DS;
+                echo Json::encode([
+                    'filelink' => $base_path . $dir . DS . $unique_name,
+                    'uid' => $model->id,
+                    'filename' => $file->name,
+                    'size' => $file->size,
+                    'maxsize' => ini_get('upload_max_filesize')
+                ]);
+            }
         } else {
-            $dir = date("Ym");
-            $this->filesystem->mkdir($this->path . $dir);
-            $unique_name = preg_replace('/\s+/', '-', date("YmdHis") . '-' . rand(0, 999999999) . '-' . $file->name);
-
-            $file->saveAs($this->path . $dir . DS . $unique_name);
-
-            /**
-             * Save DocumentModel name and order attr to database
-             */
-            $model = new DocumentModel();
-            $model->user_id = Yii::$app->user->identity->getId();
-            $model->name = $file->name;
-            $model->orginal_name = $file->name;
-            $model->unique_name = $unique_name;
-            $model->type = FileManager::FILE_TYPE_DOCUMENT;
-            $model->size = "$file->size";
-            $model->file_type = $file->type;
-            $model->description = "Dokumen " . $file->name;
-            $model->create_at = date("Y-m-d H:i:s");
-            $model->update_et = date("Y-m-d H:i:s");
-            $model->save();
-            $base_path = Yii::getAlias('@web') . DS . 'resources' . DS . 'documents' . DS;
-            echo Json::encode([
-                'filelink' => $base_path . $dir . DS . $unique_name,
-                'uid' => $model->id,
-                'filename' => $file->name,
-                'size' => $file->size,
-                'maxsize' => ini_get('upload_max_filesize')
-            ]);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
     public function actionLoadredactorfile()
     {
+        if (Yii::$app->user->can('documentloadredactorfile')) {
+            if (Yii::$app->request->isAjax) {
+                $file = DocumentModel::find();
+                $param = Yii::$app->request->getQueryParams();
+                $data = [];
+                $per_page = 6;
+                $page = isset($param["FormSearch"]['page']) ? $param["FormSearch"]['page'] : 0;
+                $key = isset($param["FormSearch"]['keyword']) ? $param["FormSearch"]['keyword'] : "";
+                $position = ($page * $per_page);
+                $query = $file->onCondition(['type' => FileManager::FILE_TYPE_DOCUMENT])
+                        ->orFilterWhere(['like', 'name', $key])
+                        ->orFilterWhere(['like', 'orginal_name', $key])
+                        ->orFilterWhere(['like', 'unique_name', $key])
+                        ->orFilterWhere(['like', 'description', $key])
+                        ->orderBy(['create_at' => SORT_DESC])
+                        ->limit($per_page)
+                        ->offset($position)
+                        ->all();
+                /** @var File $v */
+                foreach ($query as $v) {
+                    $data[] = [
+                        'filelink' => Yii::getAlias('@web') . "/resources/documents/" . date('Ym', strtotime($v->create_at)) . '/' . $v->unique_name,
+                        'title' => $v->name,
+                        'des' => $v->description,
+                        'size' => $v->size,
+                        'page' => ($page + 1),
+                        'key' => $key
+                    ];
+                }
 
-        if (Yii::$app->request->isAjax) {
-            $file = DocumentModel::find();
-            $param = Yii::$app->request->getQueryParams();
-            $data = [];
-            $per_page = 6;
-            $page = isset($param["FormSearch"]['page']) ? $param["FormSearch"]['page'] : 0;
-            $key = isset($param["FormSearch"]['keyword']) ? $param["FormSearch"]['keyword'] : "";
-            $position = ($page * $per_page);
-            $query = $file->onCondition(['type' => FileManager::FILE_TYPE_DOCUMENT])
-                    ->orFilterWhere(['like', 'name', $key])
-                    ->orFilterWhere(['like', 'orginal_name', $key])
-                    ->orFilterWhere(['like', 'unique_name', $key])
-                    ->orFilterWhere(['like', 'description', $key])
-                    ->orderBy(['create_at' => SORT_DESC])
-                    ->limit($per_page)
-                    ->offset($position)
-                    ->all();
-            /** @var File $v */
-            foreach ($query as $v) {
-                $data[] = [
-                    'filelink' => Yii::getAlias('@web') . "/resources/documents/" . date('Ym', strtotime($v->create_at)) . '/' . $v->unique_name,
-                    'title' => $v->name,
-                    'des' => $v->description,
-                    'size' => $v->size,
-                    'page' => ($page + 1),
-                    'key' => $key
-                ];
+                echo Json::encode($data);
+            } else {
+                throw new NotFoundHttpException('The requested page does not exist.');
             }
-
-            echo Json::encode($data);
         } else {
-            throw new NotFoundHttpException('The requested page does not exist.');
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
@@ -130,13 +137,17 @@ class DocumentController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new DocumentSearc;
-        $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
+        if (Yii::$app->user->can('documentindex')) {
+            $searchModel = new DocumentSearc;
+            $dataProvider = $searchModel->search(Yii::$app->request->getQueryParams());
 
-        return $this->render('index', [
-                    'dataProvider' => $dataProvider,
-                    'searchModel' => $searchModel,
-        ]);
+            return $this->render('index', [
+                        'dataProvider' => $dataProvider,
+                        'searchModel' => $searchModel,
+            ]);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
@@ -146,9 +157,13 @@ class DocumentController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-                    'model' => $this->findModel($id),
-        ]);
+        if (Yii::$app->user->can('documentview')) {
+            return $this->render('view', [
+                        'model' => $this->findModel($id),
+            ]);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
@@ -158,14 +173,18 @@ class DocumentController extends Controller
      */
     public function actionCreate()
     {
-        $model = new File;
+        if (Yii::$app->user->can('documentcreate')) {
+            $model = new File;
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('create', [
+                            'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('create', [
-                        'model' => $model,
-            ]);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
@@ -177,14 +196,18 @@ class DocumentController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (Yii::$app->user->can('documentupdate')) {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('update', [
+                            'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('update', [
-                        'model' => $model,
-            ]);
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
         }
     }
 
@@ -196,9 +219,13 @@ class DocumentController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->can('documentdelete')) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            throw new HttpException(403, 'You are not allowed to access this page', 0);
+        }
     }
 
     /**
